feat: fixed public route
