+++ src/app/layout.tsx
@@ -1,6 +1,10 @@
+"use client"
+import PrivateRoute from "@/components/PrivateRoute";
+import { checkIsPublicRoute } from "@/functions/check-is-public-route";
 import "@/style/tailwind.css";
 import type { Metadata } from "next";
 import { Work_Sans } from 'next/font/google';
+import { usePathname } from "next/navigation";
 
 const WorkSans = Work_Sans({
   subsets: ['latin'],
@@ -8,18 +12,35 @@ const WorkSans = Work_Sans({
   variable: '--font-WorkSans',
 });
 
-export const metadata: Metadata = {
+/* export const metadata: Metadata = {
   title: "Recicle Mais",
-};
+}; */
 
 export default function RootLayout({
   children,
 }: Readonly<{
   children: React.ReactNode;
 }>) {
+  const pathname = usePathname();
+
+  const isPublicPage = checkIsPublicRoute(pathname!);
+
   return (
     <html lang="pt-BR">
       <body className={`${WorkSans.variable}`}>
+      {isPublicPage && (
+        <>
+          <p>ROTA PUBLICA</p>
+          {children}
+        </>
+      )}
+
+      {!isPublicPage && (
+        <>
+          <p>ROTA PRIVADA</p>
+          <PrivateRoute>{children}</PrivateRoute>
+        </>
+      )}
           {children}
       </body>
     </html>
