+++ src/components/PrivateRoute/index.tsx
@@ -1,3 +1,4 @@
+"use client"
 import { useRouter } from "next/navigation";
 import { ReactNode, useEffect } from "react";
